Create new ModelStub class for convenience
(WIP) Create tests
for RemoveBuyersFromListingCommand
