+++ src/test/java/seedu/address/logic/commands/AddBuyersToListingCommandTest.java
@@ -8,20 +8,13 @@ import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.math.BigDecimal;
-import java.nio.file.Path;
 import java.util.Set;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Listings;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyListings;
-import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.ModelStub;
 import seedu.address.model.listing.Address;
 import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Listing;
@@ -140,148 +133,9 @@ public class AddBuyersToListingCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getListingsFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListingsFilePath(Path listingsFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListings(ReadOnlyListings listings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyListings getListings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-        @Override
-        public Person getPersonByName(Name name) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListing(Listing target, Listing editedListing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Listing getListingByName(Name name) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasListingsForSeller(Person seller) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Listing> getFilteredListingList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredListingList(Predicate<Listing> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single person with no listing.
      */
-    private class ModelStubWithPersonWithoutListing extends AddBuyersToListingCommandTest.ModelStub {
+    private class ModelStubWithPersonWithoutListing extends ModelStub {
         private final Person person;
         private final Listing listing;
         private final Listings listings = new Listings();
