+++ src/test/java/seedu/address/logic/commands/AddListingCommandTest.java
@@ -14,24 +14,19 @@ import static seedu.address.testutil.TypicalPersons.ELLE;
 import static seedu.address.testutil.TypicalPersons.FIONA;
 import static seedu.address.testutil.TypicalPersons.GEORGE;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
 import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Listings;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.ModelStub;
 import seedu.address.model.ReadOnlyListings;
-import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Listing;
 import seedu.address.model.person.Name;
@@ -216,148 +211,9 @@ public class AddListingCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getListingsFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListingsFilePath(Path listingsFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListings(ReadOnlyListings listings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasListingsForSeller(Person seller) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Listing getListingByName(Name name) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyListings getListings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-        @Override
-        public Person getPersonByName(Name name) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addListing(Listing listing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setListing(Listing target, Listing editedListing) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Listing> getFilteredListingList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredListingList(Predicate<Listing> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single listing.
      */
-    private class ModelStubWithListing extends AddListingCommandTest.ModelStub {
+    private class ModelStubWithListing extends ModelStub {
         private final Listing listing;
         private final ArrayList<Listing> listingsAdded = new ArrayList<>();
         private final ArrayList<Person> persons = new ArrayList<>();
@@ -410,7 +266,7 @@ public class AddListingCommandTest {
     /**
      * A Model stub that always accept the listing being added.
      */
-    private class ModelStubAcceptingListingAdded extends AddListingCommandTest.ModelStub {
+    private class ModelStubAcceptingListingAdded extends ModelStub {
         private final ArrayList<Listing> listingsAdded = new ArrayList<>();
         private final ArrayList<Person> persons = new ArrayList<>();
